requirements.txt
tensorflow>=2.16
tensorflow-datasets
numpy
pillow
matplotlib
scikit-learn

